implement eevee's temporal anti aliasing
From what I understood, eevee's TAA requires a static view and that no
objects inside frustum are moving (if we apply taa while the objects are
moving, they are blured -> nice motion blur per object effects by the
way)

If these conditions are true, we apply the filter. This is really
amazing.

To have a correct result in bge, I did like if taa_totsample was == 0
-> infinite temporal antialiasing.

When an object inside frustum begins to move or when the camera moves,
the filter is disabled.

The only missing thing in the code I used is that it doesn't work when
volumetrics are enabled (as I simplified eevee volumetrics code). If we
want to have it working with volumetrics, we have to implement exactly
the same method as in eevee for volumetrics but it can be a bit tricky.

please test that, this is really great :)

Also contact shadows seems to work now :)