one more 0 to avoid "blur" in special cases
Objects with physics (dynamic/rigid bodies) have always a slight
movement even when they seem static. So if an objects with physics is
inside frustum, the taa is not applied in bge.

So we have to handle this case and allow a small "movement margin" to
apply TAA filter even if the objects have a slight movement.

The problem is that when the view is scaled to see very small objects
and that these objects are moving below the threshold we choose, the
moving objects have a kind of motion blur effects caused TAA filter.

So the threshold we choose must be > physics objects constant movement
but inferior to "a certain number" which causes blur when the view is
scaled and the objects are small.

This could be set with an UI or maybe this works with FLT_MIN (didn't
tested). For now I choose a number which covers most of situations

EDIT: this seems to work with FLT_MIN so I choose FLT_MIN