add realtime lightprobes
This is sometimes a bit complex to deal with realtime because it seems
eevee has 2 phases:

1 phase when the objects are moving
1 phase when the objects are not moving

when nothing is moving (objects inside frustum/camera), some filters
are applied to improve image quality.

In this picture: http://pasteall.org/pic/show.php?id=120690 , nothing is
moving so the filters are applied (antialiasing...) but you can see that
the reflection is more pixellated than the scene.

In the scene, I apply the AA filter without restriction because I didn't
noticed performance decrease when the objects restart to move.

But for the probe texture, I limited the AA samples because the more AA
samples are applied, the more we have a performance decrease. Maybe in
this case this would be fine to free textures used each time we apply
antialiasing and create new textures because I have the impression there
is a performance leak here.

Another thing is that sometimes (In very big scenes), the probe texture
is not rendered. Don't really know why.

So this needs to be improved. Else this is very nice