UPBGE: Fix slow parent relation.
A quaternion issue caused a erronous behaviour of the slow parent
relation. This issue came from the slerp function which called
a sub function ToAngleAxis when multiplying a quaternion by a
scalar, in this function a axis vector was normalized. If the axis
was a zero vector (happen when we interpolate the same quaternions)
the normalized failed a return a vector of NaN. To fix this issue
quaternion now use a SafeNormalize instead of Normalize vector
function.

Fix issue: #639.