UPBGE: Fix custom shader UV layers.
Previously since BFBGE the UV layers through texture coord
for custom shader were wrongly managed. Every texture used
a unique UV layer which forced vertices to have always 8
UV layers for maximum 8 textures.
But in UPBGE the UV layer managed by vertices is variable
and so this behaviour is wrong as it could defines texture
coord with invalid memory offset of inexistant layers.

To solve this issue the behaviour for material attributes
is applied to texture coord: they are defined for every
textures but can use shared UV layers depending on the
UV layer used by the texture.
To get the proper layer index, in BL_Shader::GetAttrib
an iteration over all the UV layers from the mesh is proceeded
to match the UV name from the material texture.

Fix issue: #644.