UPBGE: Avoid normal recalculation for modifiers.
Previously a call to RecalcNormals was procceded in BL_ShapeDeformer
when we detected that the deformer was usign modifiers, but the update
normal function work upon m_transverts and m_transnorms which are not
updated in case of BL_ModifierDeformer and even worse doesn't contain
the same number of elements.
This ends in an invalid memory read.
To fix this issue the normals are not updated manually, this wans't
necessary as the modifiers are already giving correct normal during
conversion.