UPBGE: Cleanup mesh layers info.
Previously the layers were all sotred in the same list, so any code
reading them needed to check the type of the layer. In some code it
can be inconvenient.
A better system is to store uv and color layers info in separated
listes. In consideration the type doesn't have any reason to stay.