fix console warning
I think KX_BlenderMaterial could be used for a future API for eevee and
to create textures per material but many things are not used anymore in
there. We have to choose what to keep or what to remove. Anyway, the
call to UpdateShader was causing a message in the console, so I comment
it for now