UPBGE: Replace display array modification flag by update client/server.
Previously the display array was notify for an update thanks to a modification
flag, but different user were reading this flag and one user was setting clearing
it. This caused to not notify some user that the display array was modified.

To fix this issue the new CM_Update[Server/Client] utilities are used,
display array inherit from CM_UpdateServer and the users RAS_MeshBoundingBox,
RAS_Deformer and RAS_DisplayArrayBucket are owning a update client registered
into a display array.

But before using the update client/server mechanism, CM_UpdateClient needed
to be modified to track a integer for the update flag and not only a boolean,
thsi is used in DAB to catch any update from the DA but after a call to
GetInvalidAndClear make a check of what is modified actually.

RAS_MeshBoundingBox wans't really designed to handle update client as not
struct was used to represent and used DA in the bounding box. Similar to
RAS_Deformer, RAS_MeshBoundingBox use a DisplayArraySLot struct, this
struct store the tracked DA and it's update client.