UPBGE: Implement template intantiation switch utility.
The display array templates classes need a switch to create
the proper display array instance corresponding on the format
given.

This was proceded thanks to macros but these macro were duplicated
for regular display array and batch display array. As new
classes using format to get template paramater will be used
later a general function is implemented.

This function is named CM_InstantiateTemplateSwitch,
it receives as explicit template parameters the base class to down
cast to, the up class using templates and a tuple of all the
template parameters possible for the up class. Also the function
receive a key to select compare with the tuple element instantiated
and a list of arguments passed to the up class constructor.

For display array classes purpose, a tuple is defined in RAS_VertexFormat.h
containing all RAS_VertexFormatType possible: RAS_VertexFormatTuple.
RAS_VertexFormatType now have a comparaison operator with RAS_VertexFormat
to be working properly in CM_InstantiateTemplateSwitch when compared
with the key.