UPBGE: Replace SetProg by BindProg and UnbindProg in shaders.
Using a function like SetProg and an argument to decide to bind or
not obligate to use a condition. This condition can be totally
avoided if we the enable argument is not dynamic as in our case.

This function is split in two, BindProg and UnbindProg. For
BL_Shader the BindProg function is override non virtually as the only
caller is KX_BlenderMaterial which has a BL_Shader pointer.