UPBGE: Move interpolator into SG_Controller.
Interpolator list was a common attribut of all the classes derived from
SG_Controller, in consideration this attribut is moved into SG_Controller
and the function Update iterate over all the interpolators and call for
update.

In the same time the function UpdateSpatialData always call for parent
relation update as before all the controller subclasses were returning
false in function Update.