UPBGE: Use std::unique_ptr in KX_GameObject.
Some pointer in KX_GameObject corresponded to data owned by the
game object or inexistant. In this case std::unique_ptr could be used
to ensure the free of data.

The only drawback is the requierment of a copy constructor to duplicate
the data or set they to nullptr.

Also m_clientInfo is now not anymore a pointer as it is never nullptr
and owned by the game object.