UPBGE: Reset mouse look actuator position on negative event.
Previously when the mouse actuator was deactivated then activated
the mouse gap from the mouse position on the last activation was taken
causing a high movement. This behaviour is unconsistent with the first
activation of the mouse actuator which ignore the last mouse position
(set to -1) and then just init the mouse position for the next frame.
This commit mimic this behaviour when the actuator is deactived by setting
the previous mouse position to -1.

Fix issue: #663.