UPBGE: Remove parser and Calc function.
The parser is now useless with the removal of the logic expressions,
and so a majority of the Calc are removed too with logic bricks.

In consideration Calc was only needed for comparing properties for
EXP_ListValue lookup function count(…) and index(…). This function
is replaced with EXP_Value::Equal. In the same time property are
now nameless, indeed EXP_Value is already tracking the property names
in the property map.