UPBGE: Simplify SG_Controller interaction with SG_Node.
Previously the SG_Controller were helding a pointer to the node owning them,
but this pointer was used only in Update function called by the node.
So the member is removed and the function Update receive the node pointer
instead.

Also BL_Action use an helper function AddController to add a valid controller,
controllers without any interpolators (Empty()) are ignored.