UPBGE: Fix ray cast with xray and property filtering.
A check of prop.size() == 0 was rewritten in prop.size() instead of prop.empty().

In the same time the function CheckRayCastObject is implemented to factorise the
proprety and mask check made in RayHit and NeedRayCast.
CheckRayCastObject also add a little optimization by not getting the object collision
group when the mask accept all objects.