UPBGE: Add component scritp generation.
To ease the definition of a component a button named "Create Component" is added.
This button ask for a path module.Class and generates a script named module.py
with a component class Class inside.
This template class contains a start and update function commented
and a definition block before the class for global game engine variable.

If the file already exists an error is raised.

The previous button named Add Component is now renamed Register Component.