UPBGE: Refactor scene graph node and object.
Scenegraph entities and objects are refactored to produce a better design
without opaque type (void *).
New classes SG_Object and SG_Scene are introduced and SG_Node is modified to
link to them.

SG_Scene is the scene owning all the nodes (SG_Node) from a list of root nodes
(nodes without parent). This class is inherited by KX_Scene and have some basic
function to add or remove a root node, schedule update but also to duplicate and
destruct an object.
Every SG_Node have a pointer to its scene and most of the callbacks are useless
because of the access to a real type with appropriate functions.

SG_Object is previous "client object" of a node, this base class is inherited by
KX_GameObject and have basic functions and attributes for a scene graph such as
culling box and state and the node owning this object. Every node have a pointer
to its owned object.

SG_Node is modified in the goal to automate the detection of a root node and
register it in the scene during a parenting and unparenting, and the duplication
or destruction of owned objects. Also the parenting is restricted to one way :
the child is parented to an object. This help to not miss the link from the
parent to its new child.

AddReplicaObject and DupliGroupRecurse are now using only node replication to
duplicate an object instead of duplicating the object then the node and relink
the both. AddNodeReplicaObject is always called with a valid node as it is
called by ReplicateNodeObject.

Tested with sintel desert level, image render, and culling AABB debug.