UPBGE: Move SG_Controller entirely in BL_Action.
Previously SG_Controller were owned and updated by SG_Node. This class
is not the best place for it as SG_Controller are not always modifying spatial,
for example Light or Camera controllers.
The best place is in BL_Action as this class was setting the simulation time
and updating the controllers indirectly.

In the same time the option "child" in action actuator which enabled recursive
time update for controllers is removed. Indeed every call to SetSimulatedTime
was followed by Update for the controllers in BL_Action.