UPBGE: Use render schedules for shadow and texture renderers.
Previously schedules were only used for cameras rendering to a main off screen,
this was the goal of KX_CameraRenderData and KX_SceneRenderData. But the shadows
and renderers were still not using it. Using a similar model than
KX_CameraRenderData will remove the need of a fake camera as only the important
information for rendering will be stored such as modelview, projection, frustum…

Shadows and renderers use a common rendering function excepted that the
binding/unbinding of their texture is not the same, the new function
KX_KetsjiEngine::RenderTexture proceed the rendering of a texture, it takes as
excepted a KX_TextureRenderSchedule which contains all the information.

First of all the file KX_RenderData.h is renamed KX_RenderSchedule.h as it is
rendering scheduling information. KX_TextureRenderSchedule is now defined and
owned in a list in KX_SceneRenderSchedule. It contains some similar attributes
with KX_CameraRenderSchedule and some specific as the two functions to bind and
unbind the texture (off screen) called into KX_KetsjiEngine::RenderTexture, but
also the drawing mode and some flag to optionally render the world and update
lod (done only for renderers but not for shadows).
These schedules are generated into KX_Scene for shadow
(KX_Scene::ScheduleShadowsRender) and KX_TextureRendererManager for renderers
(KX_Scene::ScheduleTexturesRender and
KX_TextureRendererManager::ScheduleRender).

At the begining of KX_KetsjiEngine::Render after retrieving the render
schedules, all scenes are iterated and the each texture schedules are passed to
KX_KetsjiEngine::RenderTexture.

The main advantage of this technique is the common render function and the
suppression of fake cameras, the inconvenient is that cube map need 6 texture
schedules for each face. Even if that looks more complex, no slow down is
noticed.

Note than planar map projection computation changed. It retreives the scene
camera frame frustum and change the clip start/end to compute a new projection
matrix without knowing the base viewport, area or framing type.