UPBGE: Use bullet 3 and support multithreading physics.
The bullet library used by blender and game engine is update to the latest bullet 3
version. Only minor modification were requeired for the switch.

This bullet version allow multi threading for rigid body collision and constraints,
the main world for this prupose is btDiscretDynamicsWorldMt. Unfortunatly this
class doesn't manage soft bodies and to fill this gap btSoftRigidDynamicsWorldMt
which is a copy of btSoftRigidDynamics on top of btDiscretDynamicsWorldMt.
Once these modification the world is initialised with a special dispatcher and
a constraint solver pool for multi threading. The solver pool own as many regular
solver by threads available.

A compatiblity issue is introduced in KX_CharacterWrapper, bullet now supports vector
for the gravity instead of simple float, then a vector is exposed into the API.

Tests:
2700 spheres on two planes :
		Previous	Current
Begining	100ms		70ms
Stable		21ms		7.5ms

16000 point constraints on 10 cube chains.
		Previous	Current
		141ms		75ms