UPBGE: Fix FuzzyZero for very small vectors. (#664)
Previously FuzzyZero for vectors was comparing the squared length
of the vector, but as it is a squared length the value are even
more low for small values. To solve this issue the function should
compare the length directly but this is too expensive.
Instead the sum of all the absolute vector components and comparing
the sum to float epsilon.