UPBGE: Replace LA_System by a simple clock.
The only goal of this class was to provide the timing but STD clock
do the same without carring about the platform. m_system in KX_KetsjiEngine
is replaced by m_clock an instance of CM_Clock which exposes two functions
one for second time and an other for nanosecond time.

In the same time KX_TimeCategoryLogger is constructed with a reference
to the engine clock to avoid passing the time for each log start / end
function.