UPBGE: Fix for bug #746 Navmesh related
Previously when the navmesh was removed from the object this one kept
the OB_NAVMESH flag set and the engine tried without success to process
it as a navmesh.

Now OB_NAVMESH flag is removed.

Additionally, Name for button "Navmesh Clear Data" is modified to
"Remove Navmesh" to avoid possible misunderstandings.