UPBGE: Fix merging of bucket manager and text material.
All texts are using a unique material per scene and so a unique
material bucket and display array bucket. They both are owned
by the bucket manager (RAS_BucketManager). In case of libload
all the material buckets of the loaded scene are moved into the
destination scene bucket manager.

In case of text material the material bucket of the loaded was
moved but also destructed when the loaded was freed at the end
of the libload, this created an illegal memory access.

To solve this issue, the only proper way found is to discard
the material bucket dedicated to text when mergin the bucket
manager and when merging the objects, for font objects force
the recreation of the text user (RAS_TextUser) to be sure that
this one is linked to the display array bucket for texts of the
destination scene.

Tested with valgrind.