UPBGE: Refactor exit code managment.
In the interest of futur way to start a game from a memory file,
the exit information should not track only the file name but also
more data.

To do so the exit informations are standardized into the struct
KX_ExitInfo which holds the exit code and the file name for the
moment. This struct is stored into the ketsji engine and returned
by KX_KetsjiEngine::GetExitInfo.

Inside LA_Launcher::EngineNextFrame an exit code can be generated
(only for window exit or exit key) and the exit info is returned.
This same exit info is checked in LA_Launcher::EngineMainLoop
when it contains a exit code different from NO_REQUESTS this function
returns and so abort the launcher. Finally the creator of the
launcher is using this exit info in case or restart or start to get
the file name.

In the same time the restoration of cursor visibility is moved into
GPG_ghost and BL_KetsjiEmbedStart instead of the launcher.