UPBGE: Refactor material shader management.
This commit aim a refactor of shader management. It was noticed that materials
don't use the same shader at every rendering, for example shadow rendering
override material shaders by using a shader for simple shadows or VSM.
This situation was first fixed by using an override state during render to
select the proper shader, with the drawback to use common material attrbutes
such as VAO.
Later with the implementation of hardware skinning and the need of more specific
vertex attributes, this solution will be limited.

The new solution is to link multiple shaders to the same material per drawing
mode. This create a new level in render tree:

Manager (RAS_ManagerBucket):
	Material (RAS_MaterialBucket):
		Shader (RAS_IMaterialShader):
			Mesh Part (RAS_DisplayArrayBucket):
				Object (RAS_MeshSlot)

Unless other data type in this tree, shaders are not owning anything, only a
node linked to the shader is used in the render tree to contains mesh parts to
render with this shader.

First of all KX_BlenderMaterial is renamed into BL_Material based onto
RAS_IMaterial (new name for RAS_IPolyMaterial). RAS_IMaterial have a function
named GetShader(drawingMode) which is in charge of returning the proper shader
depending on the drawing mode.
The returned shader is of type RAS_IMaterialShader, a shader dedicated for
material as opposite of filters. This class have the old methods about
activating, deactivating, binding instancing from the material, they are all
pure virtual and redefined in subclasses RAS_OverrideShader, RAS_TextShader,
BL_MaterialShader and KX_materialShader.

Basically the material shader interface provides a function to bind the shader
(Activate), unbind (Deactivate), bind instancing attributes
(ActivateInstancing), update object transform when not in instancing
(ActivateMeshUser), get vertex attributes (GetAttribs) and get instancing
attributes (GetInstancingAttribs).
The material shader is dependent of the way the geometry is sent to screen
(currently normal or instancing), this is named geometry mode and used by
RAS_DisplayArrayBucket to know which render node to produce for instancing or
not, this mode is retrieved by GetGeomMode.
RAS_IMaterialShader is now a node of the render tree, to achieve it, it holds
two nodes, one for downward tree and a second for upward tree. Also a node data
is provided which only holds a pointer to the material shader.

RAS_OverrideShader is the interface for all the basic shader shared along
materials, they are used for shadow drawing but also wire drawing. As they are
shared the initialization is global and proceeded by a static constructor
RAS_OverrideShader::InitShaders and destructor RAS_OverrideShader::DeinitShaders
then get through RAS_OverrideShader::GetShader(type) where type can be black,
black instancing, vsm, vsm instancing. BL_Material::GetShader(…) is directly
using RAS_OverrideShader::GetShader for drawing mode shadow, shadow vsm and
wire.

RAS_TextShader is a dummy material shader that does not bind anything but let
text be considered as complete renderable objects.

BL_MaterialShader is the previous class BL_BlenderShader.

KX_MaterialShader is the material shader for the old BL_Shader which is now
renamed KX_Shader. This material shader inherite from KX_Shader to extend its
python API with attributes such as bindCallbacks and objectCallbacks, which
remove them from KX_2DFilter python API.

Geometry mode as explained is the way the geometry is sent to screen and so
drawing mode is the way the geometry sent to screen is painted. Saying that the
only drawing mode available are wire, textured, shadow, shadow variance. The
drawing mode RAS_RENDERER is removed because useless, and the category shadow
variance is added because it's a different way of painting than category shadow.

As the shader embed the geometry type, the construction of the render tree can
be made once by drawing mode for unsorted mesh and sorted mesh. The function
RAS_BucketManager::Renderbuckets is then simplified, to use 5 material
categories: alpha, solid, solid and alpha, alpha depth and shadow.